requirements.txt
amqp==5.3.1
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.4.0
billiard==4.2.2
celery==5.5.3
certifi==2025.8.3
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.6
dj-database-url==3.0.1
Django==5.2.6
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-environ==0.12.0
django-multiselectfield==1.0.1
django-redis==6.0.0
django-timezone-field==7.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
gunicorn==23.0.0
h11==0.16.0
idna==3.10
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
messagebird==2.2.0
msgpack==1.1.2
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
redis==5.2.1
referencing==0.37.0
requests==2.32.5
rpds-py==0.28.0
six==1.17.0
sqlparse==0.5.3
stripe==13.2.0
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.11.0
